Fix 0022 FK: person_field_override.person_id must be BIGINT
person.person_id is BIGINT, so the INT column made the foreign key malformed
(errno 150). Match the referenced type. The failed CREATE left no table, so
re-running bin/migrate.php applies the corrected migration cleanly.

+++ docs/schema.sql
@@ -160,7 +160,7 @@ CREATE TABLE assignment (
 -- here so feed imports leave it alone instead of reverting it. `field` is the
 -- golden-record column name, or 'title' for the primary assignment title.
 CREATE TABLE person_field_override (
-  person_id  INT          NOT NULL,
+  person_id  BIGINT       NOT NULL,
   field      VARCHAR(64)  NOT NULL,
   actor      VARCHAR(60)  NULL,
   note       VARCHAR(255) NULL,
